+++ test/unit/action_generics_models.dart
@@ -1,8 +1,10 @@
+library action_generics_models;
+
 import 'package:built_redux/built_redux.dart';
 import 'package:built_value/built_value.dart';
 
 part 'action_generics_models.g.dart';
-part 'action_generics_models.built_redux.dart';
+part 'action_generics_models.built_redux.g.dart';
 
 // ActionGenerics contains actions that have payload
 // of different generic types. This is mainly to test the generation of