+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 8.0.0-dev.0
+
+* add build.yaml and build.dart to comply with build_runner 0.7.x
+
+## 7.4.0
+
+* add built_redux_test_utils library - exposes an expectDispatched function for expecting a given action gets dispatched asynchronously
+* add combine to MiddlewareBuilder class - lets you combine middleware builders
+* add toString to Action class
+* add example
+* add docs
+
 ## 7.3.3
 
 * add changelog